requirements.txt
# ══════════════════════════════════════════════════════════════════════════════
# requirements.txt — GenAI Python Examples
# ──────────────────────────────────────────────────────────────────────────────
# Install all dependencies with:
#   pip install -r requirements.txt
# ══════════════════════════════════════════════════════════════════════════════


# ── Google Gemini API (Track 1) ────────────────────────────────────────────────
# Used in: GoogleAPIExample/, SummaryExampleAPI/, ChatBotExample/, PDFSummaryExample/
google-genai>=1.0

# Used in: all Track 1 and Track 1.5 modules (loads API keys from .env)
python-dotenv>=1.0


# ── Azure OpenAI / OpenAI API ──────────────────────────────────────────────────
# Used in: AzureOpenAIExample/
openai>=1.0

# ── Web Framework (chatbot.py browser UI) ─────────────────────────────────────
flask>=3.0


# ── Core ML Frameworks (Track 2) ──────────────────────────────────────────────
# Used in: test_VAE.py, test_GAN.py
tensorflow>=2.10

# Used in: test_StableDiffusion.py, test_Transformers.py
torch>=2.0


# ── Image Generation ───────────────────────────────────────────────────────────
# Used in: test_StableDiffusion.py
diffusers>=0.20
accelerate>=0.20


# ── NLP / Transformers ─────────────────────────────────────────────────────────
# Used in: test_StableDiffusion.py, test_Transformers.py
transformers>=4.30


# ── Data & Visualization ───────────────────────────────────────────────────────
# Used in: test_VAE.py, test_GAN.py, test_StableDiffusion.py
numpy>=1.21
matplotlib>=3.5

# Used in: test_Transformers.py
pandas>=1.3